+++ src/app/child-dev-project/notes/notes-related-to-entity/notes-related-to-entity.component.html
@@ -10,5 +10,8 @@
 >
 </app-entities-table>
 
-<app-custom-form-link-button [entity]="entity" [entityType]="entityCtr">
+<app-custom-form-link-button
+  [linkedEntity]="entity"
+  [formEntityType]="entityCtr"
+>
 </app-custom-form-link-button>
