+++ src/app/core/entity-details/related-entities-with-summary/related-entities-with-summary.component.html
@@ -15,5 +15,8 @@
   </ng-container>
 </div>
 
-<app-custom-form-link-button [entity]="entity" [entityType]="entityCtr">
+<app-custom-form-link-button
+  [linkedEntity]="entity"
+  [formEntityType]="entityCtr"
+>
 </app-custom-form-link-button>
