+++ src/app/core/entity-details/related-entities/related-entities.component.html
@@ -10,5 +10,8 @@
   [editable]="editable"
 ></app-entities-table>
 
-<app-custom-form-link-button [entity]="entity" [entityType]="entityCtr">
+<app-custom-form-link-button
+  [linkedEntity]="entity"
+  [formEntityType]="entityCtr"
+>
 </app-custom-form-link-button>
