+++ src/app/core/entity-details/related-time-period-entities/related-time-period-entities.component.html
@@ -16,5 +16,8 @@
 >
 </app-entities-table>
 
-<app-custom-form-link-button [entity]="entity" [entityType]="entityCtr">
+<app-custom-form-link-button
+  [linkedEntity]="entity"
+  [formEntityType]="entityCtr"
+>
 </app-custom-form-link-button>
