+++ src/app/features/public-form/custom-form-link-button/custom-form-link-button.component.html
@@ -2,16 +2,19 @@
   <button
     class="margin-top-regular"
     mat-flat-button
-    i18n
     (click)="copyLink()"
     matTooltip="This link to the `{{
       matchingCustomForm?.title
     }}` public form contains an individual ID that links submissions back to this {{
-      entity.getConstructor().ENTITY_TYPE
+      linkedEntity.getConstructor().ENTITY_TYPE
     }}."
     i18n-matTooltip
   >
-    <span i18n>Copy link to individualized public form ({{ matchingCustomForm?.title }})</span>
+    <span i18n
+      >Copy link to individualized public form ({{
+        matchingCustomForm?.title
+      }})</span
+    >
     <fa-icon class="color-accent" size="xl" icon="copy"></fa-icon>
   </button>
 }
